Move functions from science_direct_events.py into hi_l1a.py
